pyproject.toml
[tool.poetry]
name = "hw6"
version = "0.1.0"
description = ""
authors = ["Michael Kuian <michaelkuian@gmail.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
fake = "^0.8"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
